artist-assistant_ITAI4373/requirements.txt
streamlit
python-dotenv
openai
pandas
